Consolidate favorite UI in Favorite.jsx
