+++ src/Device/DeviceSelector/DeviceList/device.scss
@@ -44,16 +44,6 @@
         color: $nordic-blue;
         background-color: $white;
 
-        .favBtn {
-            width: 130px;
-            height: 100%;
-            border: 0px solid $white;
-            border-top: 1px solid $nordic-blue;
-            border-right: 1px solid $nordic-blue;
-            .star {
-                margin-top: 9px;
-            }
-        }
         .inputBtn {
             position: absolute;
             width: 130px;
