+++ src/Device/DeviceSelector/basic-device-info.scss
@@ -26,10 +26,6 @@
             flex-direction: column;
             margin-right: 20px;
             margin-left: 10px;
-
-        }
-        .favorite{
-            color: goldenrod !important ;
         }
     }
 }
