+++ src/Device/DeviceSelector/favorite.scss
@@ -1,0 +1,21 @@
+@import "../../variables";
+
+.core19-device-selector {
+    .make-favorite {
+        width: 130px;
+        height: 100%;
+        border: 0px solid $white;
+        border-top: 1px solid $nordic-blue;
+        border-right: 1px solid $nordic-blue;
+        text-transform: uppercase;
+        padding-top: 9px;
+
+        .star {
+            margin-right: 1ex;
+        }
+    }
+
+    .is-favorite {
+        color: goldenrod;
+    }
+}
